requirements.txt
gradio
transformers
torch
torchvision
scikit-image
flask
rembg
onnxruntime
pillow
numpy





